Moving mime type / size / original name detection to the core Storage implementation
