+++ lib/storage/adapters/fs/index.js
@@ -1,6 +1,4 @@
 var fs = require('fs-extra');
-var path = require('path');
-var mime = require('mime-types');
 
 function FSAdapter (options) {
 	this.options = options;
@@ -23,25 +21,6 @@ FSAdapter.prototype.getFileURL = function (value) {
 	file information spec](https://github.com/expressjs/multer#file-information)
 */
 FSAdapter.prototype.uploadFile = function (file, callback) {
-	// The path to the file is required
-	if (!file.path) return callback();
-	// Detect required information if it wasn't provided by inspecting the
-	// file stored at file.path
-	if (!file.mimetype) file.mimetype = mime(file.path);
-	if (!file.size) {
-		try {
-			var stats = fs.statSync(file.path);
-			if (!stats.isFile()) {
-				return callback(new Error(file.path + ' is not a file'));
-			}
-			file.size = stats.size;
-		} catch (e) {
-			return callback(e);
-		}
-	}
-	if (!file.originalname) {
-		file.originalname = file.name || file.path.substr(file.path.lastIndexOf(path.sep) + 1);
-	}
 	console.log('FS Adapter Uploading File:', file);
 	// TODO: Move the file into the destination path and name it correctly
 	callback(null, {
