+++ lib/storage/index.js
@@ -1,4 +1,7 @@
 var assign = require('object-assign');
+var fs = require('fs-extra');
+var path = require('path');
+var mime = require('mime-types');
 
 /**
 	The Adapter Compatibility Level is used to ensure that provided adapters
@@ -92,6 +95,35 @@ function Storage (options) {
 	public url for the file. The result is what is saved back to the field.
 */
 Storage.prototype.uploadFile = function (file, callback) {
+	// The path to the file is required
+	if (!file.path) return callback();
+
+	// Detect required information if it wasn't provided by inspecting the
+	// file stored at file.path
+	if (this.schema.filetype) {
+		if (!file.mimetype) file.mimetype = mime(file.path);
+	}
+	if (this.schema.filesize) {
+		if (!file.size) {
+			try {
+				// TODO: Improve this by removing the sync call; requires this whole
+				// code block to be more complex & call adapter.upload after the
+				// callback so I've gone with the simple statSync option for now
+				var stats = fs.statSync(file.path);
+				if (!stats.isFile()) {
+					return callback(new Error(file.path + ' is not a file'));
+				}
+				file.size = stats.size;
+			} catch (e) {
+				return callback(e);
+			}
+		}
+	}
+	if (!file.originalname) {
+		file.originalname = file.name || file.path.substr(file.path.lastIndexOf(path.sep) + 1);
+	}
+
+	// Use the adapter to upload the file
 	var adapter = this.adapter;
 	adapter.uploadFile(file, function (err, result) {
 		if (err) return callback(err);
